moonbix: Update API
- thanks to tele automation_tools / scriptvip.
